refactor(core): extract orchestrator foundation modules (1/3)
Move pure module-level helpers out of the 3953-line orchestrator.ts into five
focused modules, with zero behavior change:

- run-context.ts: RunContext/ActiveCheckpoint types, run identity/metadata
  helpers, and the token/error utilities (the dependency-free base)
- budget.ts: token/cost accounting, budget-cap enforcement, run-metrics rollup
- short-circuit.ts: isSimpleGoal / selectBestAgent
- retry.ts: executeWithRetry / computeRetryDelay
- agent-config.ts: buildAgent, default-tool-preset grant, model routing

orchestrator.ts drops to 3305 lines and re-exports isSimpleGoal, selectBestAgent,
computeRetryDelay, and executeWithRetry, so the index.ts barrel and all test
import paths are unchanged.

Pure move: extracted bodies are byte-identical to the originals (only an export
prefix added) and the kept region is byte-identical to origin/main. Verified with
npm run lint plus 1385 passing tests, zero diff under tests/, and a
dist/index.d.ts identical to the pre-split build.

First of three sequential pure-move PRs (PR2: consensus + task-execution;
PR3: coordinator + AGENTS.md).
