+++ packages/core/src/orchestrator/agent-config.ts
@@ -1,0 +1,110 @@
+/**
+ * @fileoverview Agent construction and model-routing helpers.
+ *
+ * Builds a minimal {@link Agent} with its own registry/executor, applies the
+ * orchestrator's default tool preset (default-deny grant), and resolves
+ * per-phase model-routing overrides.
+ */
+
+import type {
+  AgentConfig,
+  ModelRouteConfig,
+  ModelRoutingPolicy,
+  OrchestratorConfig,
+  Task,
+} from '../types.js'
+import { Agent } from '../agent/agent.js'
+import { ToolRegistry } from '../tool/framework.js'
+import { ToolExecutor } from '../tool/executor.js'
+import { registerBuiltInTools } from '../tool/built-in/index.js'
+
+/**
+ * Build a minimal {@link Agent} with its own fresh registry/executor.
+ * Pool workers pass `includeDelegateTool` so `delegate_to_agent` is available during `runTeam` / `runTasks`.
+ */
+export function buildAgent(
+  config: AgentConfig,
+  toolRegistration?: { readonly includeDelegateTool?: boolean },
+): Agent {
+  const registry = new ToolRegistry()
+  registerBuiltInTools(registry, toolRegistration)
+  if (config.customTools) {
+    for (const tool of config.customTools) {
+      registry.register(tool, { runtimeAdded: true })
+    }
+  }
+  const executor = new ToolExecutor(registry, {
+    ...(config.maxToolOutputChars !== undefined
+      ? { maxToolOutputChars: config.maxToolOutputChars }
+      : {}),
+  })
+  return new Agent(config, registry, executor)
+}
+
+/**
+ * Apply the orchestrator's {@link OrchestratorConfig.defaultToolPreset} as a
+ * fallback grant for an agent that declares neither `tools` nor `toolPreset`.
+ *
+ * Built-in tools are opt-in (default-deny): an agent with no grant resolves to
+ * zero built-in tools. This fills that gap when the orchestrator opts in to a
+ * default. Per-agent grants always win — the default never widens an agent that
+ * already declares `tools` or `toolPreset`.
+ */
+export function applyDefaultToolPreset(
+  config: AgentConfig,
+  defaultToolPreset: OrchestratorConfig['defaultToolPreset'],
+): AgentConfig {
+  if (
+    defaultToolPreset === undefined
+    || config.tools !== undefined
+    || config.toolPreset !== undefined
+  ) {
+    return config
+  }
+  return { ...config, toolPreset: defaultToolPreset }
+}
+
+export interface ModelRoutingSelection {
+  readonly phase: 'coordinator' | 'synthesis' | 'short-circuit' | 'worker' | 'delegated'
+  readonly agent: string
+  readonly task?: Task
+  readonly leaf?: boolean
+}
+
+export function routeMatches(
+  policy: ModelRoutingPolicy | undefined,
+  selection: ModelRoutingSelection,
+): ModelRouteConfig | undefined {
+  if (!policy) return undefined
+  const task = selection.task
+  for (const rule of policy.rules) {
+    const match = rule.match
+    if (match.phase !== undefined && match.phase !== selection.phase) continue
+    if (match.agent !== undefined && match.agent !== selection.agent) continue
+    if (match.taskRole !== undefined && match.taskRole !== task?.role) continue
+    if (match.taskPriority !== undefined && match.taskPriority !== task?.priority) continue
+    if (match.leaf !== undefined && match.leaf !== selection.leaf) continue
+    if (match.hasDependencies !== undefined && match.hasDependencies !== ((task?.dependsOn?.length ?? 0) > 0)) continue
+    return rule.route
+  }
+  return undefined
+}
+
+export function withModelRoute(config: AgentConfig, route: ModelRouteConfig | undefined): AgentConfig {
+  if (!route) return config
+  return {
+    ...config,
+    model: route.model,
+    provider: route.provider ?? config.provider,
+    baseURL: route.baseURL ?? config.baseURL,
+    apiKey: route.apiKey ?? config.apiKey,
+    region: route.region ?? config.region,
+  }
+}
+
+export function isLeafTask(task: Task, tasks: readonly Task[]): boolean {
+  for (const candidate of tasks) {
+    if (candidate.dependsOn?.includes(task.id)) return false
+  }
+  return true
+}
